reorganize corrections
